Restructure different UI modes into `duke.ui` packages
