+++ src/main/java/duke/Launcher.java
@@ -1,8 +1,7 @@
 package duke;
 
-import java.util.Scanner;
-
-import duke.ui.ConsoleUi;
+import duke.ui.console.ConsoleApplication;
+import duke.ui.graphical.GraphicalApplication;
 import javafx.application.Application;
 
 /**
@@ -16,39 +15,10 @@ public class Launcher {
      */
     public static void main(String[] args) {
         if (args.length == 1 && args[0].equals("--console")) {
-            runConsole();
+            ConsoleApplication.start();
         } else {
-            Application.launch(Main.class, args);
+            Application.launch(GraphicalApplication.class, args);
         }
     }
 
-    /**
-     * Initializes a instance of {@link Duke} in the console.
-     */
-    public static void runConsole() {
-        String filePath = "duke.txt";
-        Duke duke;
-
-        try {
-            duke = new Duke(filePath);
-        } catch (Exception e) {
-            // TODO: figure out static/non-static Ui class
-            System.out.println("Unable to initialize data file");
-            return;
-        }
-
-        Scanner inputScanner = new Scanner(System.in);
-        ConsoleUi ui = new ConsoleUi(inputScanner);
-        ui.printGreeting();
-        duke.setUi(ui);
-
-        boolean shouldExit = false;
-        while (!shouldExit) {
-            String fullCommand = ui.readCommand();
-            shouldExit = duke.executeCommand(fullCommand);
-        }
-
-        ui.printGoodbye();
-        inputScanner.close();
-    }
 }
