+++ src/main/java/duke/ui/graphical/GraphicalApplication.java
@@ -1,7 +1,8 @@
-package duke;
+package duke.ui.graphical;
 
 import java.io.IOException;
 
+import duke.Duke;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
@@ -9,9 +10,9 @@ import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
 /**
- * A GUI for Duke using FXML.
+ * Initializes the main graphical application for Duke using FXML.
  */
-public class Main extends Application {
+public class GraphicalApplication extends Application {
     private Duke duke;
 
     @Override
@@ -19,7 +20,7 @@ public class Main extends Application {
         try {
             duke = new Duke("duke.txt");
 
-            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            FXMLLoader fxmlLoader = new FXMLLoader(GraphicalApplication.class.getResource("/view/MainWindow.fxml"));
             AnchorPane ap = fxmlLoader.load();
             Scene scene = new Scene(ap);
             stage.setScene(scene);
